# openapi.yaml
openapi:  3.1.0
servers: 
  - url: http://127.0.0.1:8080
    description: Local Test server
info: 
  title: Authust API V1
  version: '1.0'
paths:
  "/api/v1/flow/executor/{flow-slug}":
    get: 
      summary: Get Flow
      operationId: get-flow-execution
      parameters:
        - $ref: "#/components/parameters/FlowSlug"
        - $ref: "#/components/parameters/Query"
      responses: 
        200: 
          description: Ok
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/FlowData"
      security: 
        - {}
        - auth: []
    post: 
      summary: Post Flow
      operationId: post-flow-execution
      parameters:
        - $ref: "#/components/parameters/FlowSlug"
        - $ref: "#/components/parameters/Query"
      responses: 
        200:
          description: Ok
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/FlowData"
      security:
        - {}
        - auth: []

  "/api/v1/auth":
    get: 
      summary: Check Auth
      operationId: check-auth
      responses: 
        200: 
          description: Ok
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/PartialUser"
        401:
          description: Unauthorized
      security:
        - {}
        - auth: []
  "/flow/-/default/{flow-designation}":
    get: 
      summary: Redirect to flow execution
      operationId: redirect-flow-default
      parameters: 
        - $ref: "#/components/parameters/FlowDesignation"
      security: 
        - {}
        - auth: []
      responses:
        302:
          description: Redirect to flow execution
          headers:
            location:
              description: Url of the flow execution
              schema: 
                type: string
components:
  securitySchemes:
    auth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    FlowSlug:
      in: path
      required: true
      name: flow-slug
      schema:
        type: string
    Query:
      in: query
      name: query
      required: false
      schema:
        type: string
    FlowDesignation:
      in: path
      required: true
      name: flow-designation
      schema: 
        $ref: "#/components/schemas/FlowDesignation"
  schemas:
    FlowData:
      $ref: model/schemas/flow-data.json
    PartialUser:
      $ref: model/schemas/partial-user.json
    FlowDesignation:
      $ref: model/schemas/flow-designation.json
    Reference:
      $ref: model/schemas/reference.json